+++ minicube-generation/CHANGES.md
@@ -1,5 +1,6 @@
 ## Changes in 0.3.5 (in development)
 * Added pinger script
+* Move configuration files to dedicated folder after minicube creation
 
 ## Changes in 0.3.4
 * Removed cloud mask from recipe